+++ komikaan/Interfaces/IStopManagerService.cs
@@ -1,0 +1,10 @@
+﻿using komikaan.Data.Models;
+
+namespace komikaan.Interfaces
+{
+    public interface IStopManagerService
+    {
+        Task StartAsync(CancellationToken cancellationToken);
+        Task<IEnumerable<SimpleStop>> GetAllStopsAsync();
+    }
+}