+++ komikaan/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.6.0] - 08-03-2024
+### Added
+- Stop Manager to merge related stops of different suppliers
+- A simple mapping system for id mappings of different suppliers
+- SimpleStop model
+- Added Direciton, LineName and Operator to SimpleRoutePart
+### Changed
+- Mark a route as "maybe" if:
+    - There are routes cancelled for no reason
+- Mark a route as "Unknown" if:
+    - No routes can be found
+- DisruptionController now checks all data suppliers for routes
+
 [1.5.2] - 05-03-2024
 ### Added
 - URL if available for a disruption from NS